Edit code structures

+++ gulpfile.js
@@ -7,6 +7,7 @@ import debug from 'gulp-debug';
 import autoprefixer from 'gulp-autoprefixer';
 import * as sass from 'sass'
 import gsass from 'gulp-sass';
+import {pathExistsSync} from 'path-exists';
 import connectphp from 'gulp-connect-php';
 import browserSync from 'browser-sync';
 
@@ -39,8 +40,11 @@ const config = {
     }
 }
 
-const do_clean = () => {
-    return gulp.src([config.build + '/*'], { read: false }).pipe(rimraf({ force: true }));
+const do_clean = (done) => {
+    if(pathExistsSync(config.build)) {
+        return gulp.src([config.build + '/*'], { read: false }).pipe(rimraf({ force: true }));
+    }
+    done();
 }
 
 const do_browserSync = (done) => {
